Reorganize tests Make test files smaller (sperate tests for Vector2, Vector3, lists, Matrices and so on...) Put tests into folders matching their library It allows also a better overview where we need to improve our test coverage in the future
