+++ test/console_test_harness.dart
@@ -1,89 +1,66 @@
 library console_test_harness;
 
 import 'package:unittest/unittest.dart';
-import 'vector_math_test.dart';
+import 'src/vector_math/test_quaternion.dart' as test_quaternion;
+import 'src/vector_math/test_quad.dart' as test_quad;
+import 'src/vector_math/test_matrix2.dart' as test_matrix2;
+import 'src/vector_math/test_matrix3.dart' as test_matrix3;
+import 'src/vector_math/test_matrix4.dart' as test_matrix4;
+import 'src/vector_math/test_colors.dart' as test_colors;
+import 'src/vector_math/test_noise.dart' as test_noise;
+import 'src/vector_math/test_vector2.dart' as test_vector2;
+import 'src/vector_math/test_vector3.dart' as test_vector3;
+import 'src/vector_math/test_vector4.dart' as test_vector4;
+import 'src/vector_math/test_aabb2.dart' as test_aabb2;
+import 'src/vector_math/test_aabb3.dart' as test_aabb3;
+import 'src/vector_math/test_obb3.dart' as test_obb3;
+import 'src/vector_math/test_opengl.dart' as test_opengl;
+import 'src/vector_math/test_utilities.dart' as test_utilities;
+import 'src/vector_math/test_sphere.dart' as test_sphere;
+import 'src/vector_math/test_ray.dart' as test_ray;
+import 'src/vector_math/test_triangle.dart' as test_triangle;
+import 'src/vector_math/test_plane.dart' as test_plane;
+import 'src/vector_math/test_frustum.dart' as test_frustum;
+import 'src/vector_math_geometry/test_geometry.dart' as test_geometry;
+import 'src/vector_math_list/test_vector_list.dart' as test_vector_list;
 
 void main() {
   groupSep = ' - ';
 
-  group('Quadernion', () {
-    QuadTest qt = new QuadTest();
-    qt.run();
+  group('vector_math', () {
+    test_aabb2.main();
+    test_aabb3.main();
+    test_colors.main();
+    test_frustum.main();
+    test_matrix2.main();
+    test_matrix3.main();
+    test_matrix4.maib();
+    test_noise.main();
+    test_obb3.main();
+    test_opengl.main();
+    test_plane.main();
+    test_quad.main();
+    test_quaternion.main();
+    test_ray.main();
+    test_sphere.main();
+    test_triangle.main();
+    test_utilities.main();
+    test_vector2.main();
+    test_vector3.main();
+    test_vector4.main();
   });
 
-  group('Quaternion', () {
-    QuaternionTest qt = new QuaternionTest();
-    qt.run();
+  group('vector_math_64', () {
   });
 
-  group('Matrix', () {
-    MatrixTest mt = new MatrixTest();
-    mt.run();
+  group('vector_math_geometry', () {
+    test_geometry.main();
   });
 
-  group('Vector', () {
-    VectorTest vt = new VectorTest();
-    vt.run();
+  group('vector_math_list', () {
+    test_vector_list.main();
   });
 
-  group('OBB', () {
-    ObbTest ot = new ObbTest();
-    ot.run();
-  });
-
-  group('OpenGL', () {
-    OpenGLMatrixTest omt = new OpenGLMatrixTest();
-    omt.run();
-  });
-
-  group('AABB', () {
-    AabbTest at = new AabbTest();
-    at.run();
-  });
-
-  group('Colors', () {
-    ColorsTest ct = new ColorsTest();
-    ct.run();
-  });
-
-  group('Utilities', () {
-    var ut = new UtilitiesTest();
-    ut.run();
-  });
-
-  group('Geometry', () {
-    GeometryTest gt = new GeometryTest();
-    gt.run();
-  });
-
-  group('Sphere', () {
-    SphereTest st = new SphereTest();
-    st.run();
-  });
-
-  group('Ray', () {
-    RayTest rt = new RayTest();
-    rt.run();
-  });
-
-  group('Triangle', () {
-    TriangleTest tt = new TriangleTest();
-    tt.run();
-  });
-
-  group('Plane', () {
-    PlaneTest pt = new PlaneTest();
-    pt.run();
-  });
-
-  group('Frustum', () {
-    FrustumTest ft = new FrustumTest();
-    ft.run();
-  });
-
-  group('Noise', () {
-    NoiseTest nt = new NoiseTest();
-    nt.run();
+  group('vector_math_operations', () {
   });
 }
-
