+++ test/src/vector_math/test_obb3.dart
@@ -1,0 +1,8 @@
+library test_obb3;
+
+import 'package:unittest/unittest.dart';
+
+void main() {
+  group('Obb3', () {
+  });
+}
