--- test/test_obb.dart
@@ -1,12 +1,0 @@
-part of vector_math_test;
-
-class ObbTest extends BaseTest {
-
-  void testX() {
-
-  }
-
-  void run() {
-    test('X', testX);
-  }
-}
