--- test/vector_math_test.dart
@@ -1,26 +1,0 @@
-library vector_math_test;
-
-import 'package:unittest/unittest.dart';
-import 'dart:math' as Math;
-import 'dart:typed_data';
-import 'package:vector_math/vector_math.dart';
-import 'package:vector_math/vector_math_lists.dart';
-import 'package:vector_math/vector_math_geometry.dart';
-
-part 'base_test.dart';
-part 'test_quaternion.dart';
-part 'test_quad.dart';
-part 'test_matrix.dart';
-part 'test_colors.dart';
-part 'test_noise.dart';
-part 'test_vector.dart';
-part 'test_aabb.dart';
-part 'test_obb.dart';
-part 'test_opengl_matrix.dart';
-part 'test_geometry.dart';
-part 'test_utilities.dart';
-part 'test_sphere.dart';
-part 'test_ray.dart';
-part 'test_triangle.dart';
-part 'test_plane.dart';
-part 'test_frustum.dart';
